refactor(nodes): consolidate PocketbaseHttp functions under Common directory
Streamlined PocketbaseHttp node by moving shared functionality to a Common directory.

- Extracted `GenericFunctions`, `LoadOptions`, `PocketbaseAuth`, and `RequestBodyFunctions` modules.
- Updated imports in `PocketbaseHttp.node.ts` to use new paths.

This improves modularity and facilitates code reuse.
