+++ nodes/Common/GenericFunctions.ts
